Update and refactor git branch aware functions.
4 gitbranch_aware functions for saving files from matlab with a
automatic naming convention where the name of the currently checked out
branch is added as a suffix to the filename. The functions
add_filename_suffix.m and git_branch_name_suffix.m handle the finding of
the checked out branch name and appending the suffix. The 4
gitbranch_aware functions are wrappers for save(), saveas(), load() and
matfile().