Improve hrf_conv.m, work on internal_reliability.m
I discovered that hrf_conv.m was not handling design matrices given in
volume resolution - specifically it was not up sampling to millisecond
resolition before applying a millisecond resolution hrf to it... now it
does that.

Also I have the internal_reliability function working for pervoxel and
mask. I need to add the ROI average functionality and generally make
sure it's all good.